create arhitecture
