+++ lib/add_product/add_product.dart
@@ -1,0 +1,1 @@
+export 'page/add_product_page.dart';
