+++ lib/home/home.dart
@@ -1,0 +1,1 @@
+export 'page/home_page.dart';
