+++ lib/home/page/home_page.dart
@@ -1,0 +1,15 @@
+import 'package:flutter/material.dart';
+
+class HomePage extends StatelessWidget {
+  const HomePage({Key? key}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(
+        actions: [],
+        title: Text("HomePage"),
+      ),
+    );
+  }
+}
